requirements.txt
flask
flask-cors
python-dotenv
pyfiglet
colorama
inquirer
yaspin